+++ changelog.md
@@ -3,6 +3,10 @@
 
 **pycite 0.1.1**
 
+* Explicitly set an HTML parser 
+
+* Initial tests 
+
 * Volumes no longer have the leading "v" attached. 
 
 * Added `split_authors` a simple method to clean and abbreviate author names.